[Agent] move loader mock helper
