+++ tests/common/mockFactories/index.js
@@ -13,6 +13,8 @@ export {
   createMockAIPromptPipeline,
 } from './coreServices.js';
 
+export { createLoaderMocks } from './loaders.js';
+
 export { createRuleTestDataRegistry } from './entities.js';
 
 export { createMockContainerWithRegistration } from './container.js';
