+++ tests/common/mockFactories/loaders.js
@@ -109,3 +109,21 @@ export const createMockModVersionValidator = () => {
   fn.validate = fn;
   return fn;
 };
+
+/**
+ * Generates mock loader objects for the given content types.
+ *
+ * @description Iterates over the supplied type names and returns an
+ *   object containing `{ mock${type}Loader: createMockContentLoader() }`
+ *   entry for each.
+ * @param {string[]} types - Content loader type names.
+ * @returns {Record<string, { loadItemsForMod: jest.Mock }>} Mapping of
+ *   mock loader names to loader mocks.
+ */
+export function createLoaderMocks(types) {
+  const loaders = {};
+  for (const type of types) {
+    loaders[`mock${type}Loader`] = createMockContentLoader();
+  }
+  return loaders;
+}
